Added code to produce expression csv's from gtf's.
Also fixed a small bug in transcript parsing.

+++ file_types/gtf_file.py
@@ -38,7 +38,7 @@ def parse_gff_line( line, fix_chrm=True ):
         data[5] = 0
     else:
         try:
-            data[5] = int( data[5] )
+            data[5] = float(data[5])
         except ValueError:
             return None
